refactor: split timing engine tests
